Add expected behavior for NoOpAstVisitor for Body/Paragraph
While it's a "no op" visitor, the nodes that have children, for a
reasonable person, would expect to traverse the children of the node
right?

Like, if you declared a CollectTextNodes visitor and only implemented
the text node visit, then you would reasonably want to not re-implement
traversals of every potential parent element right?

I'll rename the class probably.